refactor(pipeline): extract pure quality-gate/normalization helpers
First behavior-preserving slice of the 2.6k-line pipeline.rs god file: move
the pure, well-tested helpers (recording_gate_rms, is_transcription_hallucination,
preview_text, normalize_transcription_math_artifacts) and the MIN_RECORDING_*
thresholds into a new pipeline/gates.rs submodule, re-imported into pipeline.rs
so all call sites and the test module are unchanged. No logic changes.
